code review4
Lines 237-240 could be replaced with the following vectorised operation:

x_ = Xsig_pred_ * weights_;
Although array programming is not native to C++, Eigen does provide these operations for matrix and vector calculations. Here is a helpful description of array programming.
https://en.wikipedia.org/wiki/Array_programming
Some developers may consider array programming to be less readable and avoid using it even if it is more efficient. It's fine if you agree with this, but I wanted to bring it up in case you weren't aware of this approach.

There are other opportunities to use vectorised operations in your code, but I'll leave it as an exercise for you to refactor your project if you wish to.