code review6
Every time UpdateLidar() is called, R is being re-initialised to the same matrix. It would be more efficient if the lidar and radar measurement covariance matrices were initialised once in the UKF() constructor and simply referenced as a property of the UKF (perhaps R_laser_ and R_radar_ to match the property naming conventions already used).

NB: This comment applies to the R measurement covariance matrix initialisation in both radar and lidar methods.