code review7
Since the code in this loop can't make any assumptions about the values of px and py it would be best to add defensive logic here to avoid dividing by 0. It is also worth being aware that C++ will return a value for atan2(0, 0), but the way the atan2 function is defined means atan2(0, 0) is undefined. I recommend checking out the wikipedia page on atan2 to find out more. It is up to you if you wish to add a special case for this, but I wanted to bring it to your attention just in case you weren't aware of this aspect of the function.
https://en.wikipedia.org/wiki/Atan2