code review8
The process of normalising an angle appears multiple times in your code. It would be great to see this repetition refactored into helper functions (perhaps kept in the Tools class).

This refactoring would allow you to easily swap out the normalisation code for a different process that doesn't use while loops (such as using fmod or atan2).